myproject/urls.py

import debug_toolbar
from django.conf import settings
from django.conf.urls.static import static
from django.contrib import admin
from django.urls import include, path

from myproject.views import main_view_async

urlpatterns = [
    path("admin/", admin.site.urls),
    path("", include("users.urls")),
    path('async/', main_view_async, name='async-main-view'),
    path('__debug__/', include(debug_toolbar.urls)),
]

if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
