requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.6
django-crispy-forms==1.14.0
django-environ==0.9.0
django-storages==1.13.2
django_debug_toolbar==3.8.1
google-api-core==2.10.2
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.12.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.48.2
h11==0.14.0
idna==3.4
Pillow==9.4.0
proto-plus==1.22.2
protobuf==3.20.3
psycopg2==2.9.5
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
uvicorn==0.20.0
